requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.16.4
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyDrive==1.3.1
pygame==2.6.0
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
SpeechRecognition==3.10.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
pyaudio==0.2.14
setuptools==68.0.0
openai==1.37.0
